requirements.txt
APScheduler==3.6.3
astroid==2.5.6
certifi==2020.12.5
isort==5.8.0
lazy-object-proxy==1.6.0
mccabe==0.6.1
pylint==2.8.2
python-telegram-bot==13.5
pytz==2021.1
six==1.16.0
toml==0.10.2
tornado==6.1
tzlocal==2.1
wrapt==1.12.1
emoji==1.2.0